# Migrate renorm to ATen (CPU and CUDA)