--- a/aten/src/ATen/native/Normalization.h
+++ b/aten/src/ATen/native/Normalization.h
@@ -1,11 +1,9 @@
 #pragma once
 
+#include <ATen/TensorIterator.h>
 #include <ATen/native/DispatchStub.h>
 
 namespace at {
-
-class TensorIteratorBase;
-
 namespace native {
 
 using renorm_scale_factor_fn = void (*) (TensorIteratorBase& iter, double maxnorm);