--- a/aten/src/ATen/native/cuda/DistanceKernel.cu
+++ b/aten/src/ATen/native/cuda/DistanceKernel.cu
@@ -1,5 +1,7 @@
 #include <ATen/ATen.h>
 #include <ATen/cuda/Exceptions.h>
+#include <ATen/cuda/DeviceUtils.cuh>
+#include <ATen/cuda/CUDAContext.h>
 #include <THC/THCTensorMathReduce.cuh>
 #include <math.h>